requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Django==4.2.19
django-cors-headers==4.7.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
